+++ lib/batch.dart
@@ -23,6 +23,7 @@ export 'package:batch/src/log/log_level.dart';
 export 'package:batch/src/log/input_log_event.dart';
 export 'package:batch/src/log/output_log_event.dart';
 export 'package:batch/src/log/filter/log_filter.dart';
-export 'package:batch/src/log/filter/default_log_filter.dart';
 export 'package:batch/src/log/output/console_log_output.dart';
 export 'package:batch/src/log/output/file_log_output.dart';
+export 'package:batch/src/log/color/log_color.dart';
+export 'package:batch/src/log/color/console_color.dart' show ConsoleColor;