+++ lib/src/log/log_configuration.dart
@@ -3,6 +3,7 @@
 // BSD-style license that can be found in the LICENSE file.
 
 // Project imports:
+import 'package:batch/src/log/color/log_color.dart';
 import 'package:batch/src/log/filter/log_filter.dart';
 import 'package:batch/src/log/log_level.dart';
 import 'package:batch/src/log/output/log_output.dart';
@@ -15,6 +16,7 @@ class LogConfiguration {
     this.filter,
     this.printer,
     this.output,
+    this.color,
     this.printLog = true,
   });
 
@@ -29,6 +31,9 @@ class LogConfiguration {
 
   /// The custom log output
   final LogOutput? output;
+
+  /// The custom log color
+  final LogColor? color;
 
   /// The flag represents this logger should print log or not
   final bool printLog;